Delete some unneded things, move parseComponentName to utils
