--- scripts/templates/preview.html.tmpl
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <title><%= name %></title>
-  <link rel="stylesheet" href="<%= name %>.css">
-</head>
-<body>
-  <%= body %>
-  <% if(hasJS) { %><script src="<%= name %>.js"></script><% } %>
-</body>
-</html>
