+++ scripts/utils/parseComponentName.js
